% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate simulated data that meets the parallel trends assumption}
\usage{
generate_data(
  N,
  Tt,
  Beta,
  potential_outcomes = FALSE,
  ylink = "rnorm_identity",
  binomial_n = 1,
  long = FALSE
)
}
\arguments{
\item{N}{int. Number of independent observations (i.e. units)}

\item{Tt}{int. Number of periods, minus 1. I.e. there are Tt + 1 periods.}

\item{Beta}{list of length 5. Output of generate_parameters().}

\item{potential_outcomes}{logical. Should outcomes and covariates be generated with exposure set to 0 at all times?}

\item{ylink}{chr. Generalized linear model distribution and link specification for y. Currently either "rnorm_identity" or "rbinom_logit"}

\item{binomial_n}{int length N. Defaults to all 1's. If ylink is rbinom_logit, you can optionally pass a vector of group sizes to generate aggregate binomial data. In this case, treatments and covariates will be constant at the group level for a given time period.}

\item{long}{lgl. Should the returned dataset be wide (one row per unit, FALSE), or long (Tt+1 rows per unit, TRUE) ?}
}
\value{
Data frame. If long=FALSE, resulting data frame has N rows and (Tt+1)4 + 2 columns - 'uid' is a unique identifier, 'U0' is an 'unmeasured' baseline covariate, L\{t\},A\{t\},Y\{t\} are covariates, exposures, and outcomes, respectively. Otherwise, N(Tt+1) rows and 7 columns. If binomial_n != 1, an additional column binomial_n is also included.
}
\description{
Generate simulated data that meets the parallel trends assumption
}
\examples{
Beta = generate_parameters(Tt=2)
df = generate_data(100, 2, Beta)
}
