% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{ice}
\alias{ice}
\title{Iterated conditional expectation (ICE) estimator}
\usage{
ice(
  df_obs,
  df_interv,
  inside_formula_t,
  inside_formula_tmin1,
  outside_formula,
  Tt,
  t_col,
  n_nested = Tt,
  inside_family = "gaussian",
  pt_link_fun = NULL,
  binomial_n = NULL,
  tibble = TRUE,
  models = TRUE,
  long = FALSE,
  suppress_rank_warnings = FALSE
)
}
\arguments{
\item{df_obs}{Data frame with one row per individual-period (if \code{long=TRUE}) or one row per individual (if \code{long=FALSE})}

\item{df_interv}{Data frame with same dimensions as \code{df_obs}, but with exposure variables set to intervened values}

\item{inside_formula_t}{chr, right-hand-side formula for inside model for Yt}

\item{inside_formula_tmin1}{chr, right-hand-side formula for inside model for Yt-1}

\item{outside_formula}{chr, right-hand-side formula for outside models}

\item{Tt}{int. max periods}

\item{t_col}{integer vector of length equal to \verb{nrow(df_obs).} Column denoting periods; only needed if \code{long=TRUE}.}

\item{n_nested}{int. How many nested expectations should be estimated, starting from the innermost (=0) to the outermost (=Tt)?}

\item{inside_family}{stats::family object or string referring to one, as in \code{glm}.}

\item{pt_link_fun}{function. The scale on which parallel trends is assumed (e.g., \code{qlogis} for logit scale). Default \code{NULL} for untransformed scale.}

\item{binomial_n}{int length nrow(data). Group sizes for binomial aggregate data.}

\item{tibble}{logical. return results as a tibble (TRUE) or vector (FALSE)?}

\item{models}{lgl. Return all models as an attribute?}

\item{long}{lgl. Is df_obs wide (\code{FALSE}, default) or long (\code{TRUE}) format?}
}
\value{

}
\description{
Iterated conditional expectation (ICE) estimator
}
