% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{tmle}
\alias{tmle}
\title{Targeted maximum likeilhood estimators (TMLE)}
\usage{
tmle(
  df_obs,
  df_interv,
  den_formula,
  inside_formula_t,
  inside_formula_tmin1,
  outside_formula,
  Tt,
  t_col,
  id,
  n_nested = Tt,
  den_family = "binomial",
  inside_family = "gaussian",
  pt_link_fun = NULL,
  binomial_n = NULL,
  tibble = TRUE,
  models = TRUE,
  long = FALSE,
  suppress_rank_warnings = FALSE
)
}
\arguments{
\item{df_obs}{Data frame with one row per individual-period (if \code{long=TRUE}) or one row per individual (if \code{long=FALSE})}

\item{df_interv}{Data frame with same dimensions as \code{df_obs}, but with exposure variables set to intervened values}

\item{den_formula}{chr. glue-style formula for denominator model(s)}

\item{inside_formula_t}{chr, glue-style right-hand-side formula for inside model for Yt}

\item{inside_formula_tmin1}{chr, glue-style right-hand-side formula for inside model for Yt-1}

\item{outside_formula}{chr, glue-style right-hand-side formula for outside models}

\item{Tt}{int. max periods}

\item{t_col}{integer vector of length equal to \verb{nrow(df_obs).} Column denoting periods; only needed if \code{long=TRUE}.}

\item{id}{chr. name of column in \code{df_obs} corresponding to unit identifier (only need it \code{long=TRUE})}

\item{n_nested}{int. How many nested expectations should be estimated, starting from the innermost (=0) to the outermost (=Tt)?}

\item{den_family}{stats::family object or string referring to one, as in \code{glm}, for denominator model(s).}

\item{inside_family}{stats::family object or string referring to one, as in \code{glm}, for innnermost model(s).}

\item{pt_link_fun}{function. The scale on which parallel trends is assumed (e.g., \code{qlogis} for logit scale). Default \code{NULL} for untransformed (identity) scale.}

\item{binomial_n}{int length nrow(data). Group sizes for binomial aggregate data.}

\item{tibble}{logical. return results as a tibble (TRUE) or vector (FALSE)?}

\item{models}{lgl. Return all models as an attribute?}

\item{long}{lgl. Is df_obs wide (\code{FALSE}, default) or long (\code{TRUE}) format?}

\item{suppress_rank_warnings}{lgl. Rank deficient models are often expected in this setting. Option to turn off warning 'prediction from a rank-deficient fit may be misleading'}
}
\value{
tibble with \code{Tt} rows and 2 columns. Column \code{estimate} contains estimates of counterfactual trends from t-1 to t. I.e., these are
estimates of \eqn{g\{E[Y_t(\bar a^*)]\} - g\{E[Y_{t-1}(\bar a^*)]\}}, where \eqn{g\{\cdot\}} is the parallel trends link function specified
by \code{pt_link_fun}.
}
\description{
Functions to estimate the parallel trends g-formula using double robust targeted maximum likelihood estimators.
}
\examples{
Tt = 3
N = 100
Beta = generate_parameters(Tt=Tt)
df_obs = generate_data(N, Tt, Beta)
df_interv = df_obs \%>\% dplyr::mutate(A1=0, A2=0, A3=0)
tmle(df_obs=df_obs,
    df_interv = df_interv,
    den_formula = 'A{t}~A{t-1}*(W1{t}+W2{t}+I(W2{t}^2))',
    inside_formula_t = '~A{t}*(W1{t}+W2{t}+I(W2{t}^2))',
    inside_formula_tmin1 = '~A{t}*(W1{t-1}+W2{t-1}+I(W2{t-1}^2))',
    outside_formula = '~A{k}*(W1{k}+W2{k}+I(W2{k}^2))',
    Tt=Tt)
}
